refactor: Refactored DBNet
